Use pflag for config
Changes to pflag from flag, moves most of the flag parsing and
validation logic to an options file. This now uses some of the same
conventions and patterns found in the kubernetes commands for
configuration. These patterns allow configuration to expand as needed,
allow reuse of configuration groupings, and will make configuration from
environment a little bit easier.